+++ src/index.ts
@@ -33,6 +33,7 @@ export type {
   VextValidator,
   VextConfig,
   VextUserConfig,
+  VextMiddlewareDecl,
   VextMiddlewareConfig,
   VextCorsConfig,
   VextRateLimitConfig,